[ACS-12257] simplify search_after into /search endpoint
