--- remote-api/src/main/resources/alfresco/templates/publicapi/org/alfresco/api/DeepSearchApiWebscript.post.desc.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<webscript>
-  <shortname>Handles POST for the deep pagination Search api</shortname>
-  <description>Cursor-based deep pagination search using Elasticsearch search_after</description>
-  <url>/public/search/versions/1/deep-search</url>
-  <authentication>user</authentication>
-  <transaction allow="readonly">required</transaction>
-  <format default="json">argument</format>
-  <lifecycle>public_api</lifecycle>
-</webscript>
